Remove bou.ke/monkey dependency
You mostly can't monkey patch the way you can in dynamic languages. This usually
leads to overcomplicating the test, limitations, and possible errors. For
example, in the current implementation, a subtle logical error was made, leading
to the fact that the test always passed. And in general this is not Go-way. To
do the same sort of thing, the idiomatic way in Go is to have interfaces that
you switch out. This is basically just dependency injection. However, it seems
that in these two particular cases it was enough to apply the setUp/tearDown
pattern.

In addition, `filepath.Walk` in `filekit.filepath.FindDir` was replaced by
`filepath.WalkDir` due to `Walk` is less efficient than `WalkDir`, introduced in
Go 1.16

Fixes #6